# Conflicts:
#	src/main/java/com/fivepigs/app/dao/UserDao.java

+++ src/main/java/com/fivepigs/app/web/customer/ProductServlet.java
@@ -111,7 +111,7 @@ public class ProductServlet extends HttpServlet {
         return detail.getCreatedAt();
     }
 
-    private Integer resolveFileSize(Software detail) {
+    private Long resolveFileSize(Software detail) {
         SoftwareVersion version = detail.getSoftwareVersion();
         return version == null ? null : version.getFileSize();
     }
@@ -143,7 +143,7 @@ public class ProductServlet extends HttpServlet {
         return dateTime.format(DATE_FORMAT);
     }
 
-    private String formatFileSize(Integer fileSize) {
+    private String formatFileSize(Long fileSize) {
         if (fileSize == null || fileSize <= 0) {
             return "Updating";
         }
